requirements-dev.txt
-r requirements.txt
black==23.12.1
flake8==6.1.0
django-debug-toolbar==4.1.2
pytest==7.4.3
pytest-django==4.7.0
